+++ components/Carbon.js
@@ -1,4 +1,5 @@
 import React, { PureComponent } from 'react'
+import * as hljs from 'highlight.js'
 import Spinner from 'react-spinner'
 import ResizeObserver from 'resize-observer-polyfill'
 import debounce from 'lodash.debounce'
@@ -50,20 +51,13 @@ class Carbon extends PureComponent {
     this.props.updateTitleBar(newTitle)
   }
 
-  async getHighlightLang(code = '') {
-    if (!this.hljs) {
-      this.hljs = await import('highlight.js')
-    }
-    return this.hljs.highlightAuto(code).language
-  }
-
   handleLanguageChange = debounce(
-    async (newCode, config) => {
+    (newCode, config) => {
       const props = (config && config.customProps) || this.props
 
       if (props.config.language === 'auto') {
         // try to set the language
-        const detectedLanguage = await this.getHighlightLang(newCode)
+        const detectedLanguage = hljs.highlightAuto(newCode).language
         const languageMode =
           LANGUAGE_MODE_HASH[detectedLanguage] || LANGUAGE_NAME_HASH[detectedLanguage]
