+++ components/Editor.js
@@ -6,14 +6,13 @@ import domtoimage from 'dom-to-image'
 import ReadFileDropContainer, { DATA_URL, TEXT } from 'dropperx'
 
 // Ours
-import Page from '../components/Page'
-import Button from '../components/Button'
-import Dropdown from '../components/Dropdown'
-import BackgroundSelect from '../components/BackgroundSelect'
-import Settings from '../components/Settings'
-import Toolbar from '../components/Toolbar'
-import Overlay from '../components/Overlay'
-import Carbon from '../components/Carbon'
+import Button from './Button'
+import Dropdown from './Dropdown'
+import BackgroundSelect from './BackgroundSelect'
+import Settings from './Settings'
+import Toolbar from './Toolbar'
+import Overlay from './Overlay'
+import Carbon from './Carbon'
 import api from '../lib/api'
 import {
   THEMES,
@@ -29,8 +28,8 @@ import {
   DEFAULT_CODE,
   DEFAULT_SETTINGS
 } from '../lib/constants'
-import { getQueryStringState, updateQueryString, serializeState } from '../lib/routing'
-import { getState, saveState } from '../lib/util'
+import { serializeState } from '../lib/routing'
+import { getState } from '../lib/util'
 
 const saveButtonOptions = {
   button: true,
@@ -40,30 +39,13 @@ const saveButtonOptions = {
 }
 
 class Editor extends React.Component {
-  static async getInitialProps({ asPath, query }) {
-    const path = removeQueryString(asPath.split('/').pop())
-    const queryParams = getQueryStringState(query)
-    const initialState = Object.keys(queryParams).length ? queryParams : null
-    try {
-      // TODO fix this hack
-      if (path.length >= 19 && path.indexOf('.') === -1) {
-        const content = await api.getGist(path)
-        return { content, initialState }
-      }
-    } catch (e) {
-      console.log(e)
-    }
-    return { initialState }
-  }
-
   constructor(props) {
     super(props)
     this.state = Object.assign(
       {
         ...DEFAULT_SETTINGS,
         uploading: false,
-        code: props.content,
-        _initialState: this.props.initialState
+        code: props.content
       },
       this.props.initialState
     )
@@ -84,7 +66,7 @@ class Editor extends React.Component {
 
   componentDidMount() {
     // Load from localStorage instead of query params
-    if (!this.state._initialState) {
+    if (!this.props.initialState) {
       const state = getState(localStorage)
       if (state) {
         this.setState(state)
@@ -93,12 +75,7 @@ class Editor extends React.Component {
   }
 
   componentDidUpdate() {
-    updateQueryString(this.state)
-    const s = { ...this.state }
-    delete s.code
-    delete s.backgroundImage
-    delete s.backgroundImageSelection
-    saveState(localStorage, s)
+    this.props.onUpdate(this.state)
   }
 
   getCarbonImage({ format, type } = { format: 'png' }) {
@@ -204,7 +181,7 @@ class Editor extends React.Component {
 
   render() {
     return (
-      <Page enableHeroText={true}>
+      <React.Fragment>
         <div id="editor">
           <Toolbar>
             <Dropdown
@@ -279,17 +256,9 @@ class Editor extends React.Component {
             }
           `}
         </style>
-      </Page>
+      </React.Fragment>
     )
   }
-}
-
-function removeQueryString(str) {
-  const qI = str.indexOf('?')
-  return (qI >= 0 ? str.substr(0, qI) : str)
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;')
-    .replace(/\//g, '&#x2F;')
 }
 
 function isImage(file) {
@@ -301,6 +270,10 @@ function readAs(file) {
     return DATA_URL
   }
   return TEXT
+}
+
+Editor.defaultProps = {
+  onUpdate: () => {}
 }
 
 export default DragDropContext(HTML5Backend)(Editor)
